lyrics_server/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.6
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
RapidFuzz>=3.11.0
requests==2.32.3
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
syncedlyrics==1.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
yt-dlp==2025.1.12
python-dotenv>=1.0.0
isodate>=0.6.1
pocketbase>=0.15.0
